feat: 로직 버그 수정 + 종목 내비게이션 + 차트 수정 + 랭킹/상점 실데이터 전환
- 캔들스틱 OHLC long→double 변환 (저가 코인 차트 깨짐 수정)
- TradingChart 가격대별 동적 priceFormat 적용
- 포트폴리오/대시보드에서 종목 클릭 시 해당 종목 TradePage로 이동
- TradePage URL 파라미터(?code=&type=) 기반 종목 선택
- StrategyService/QuantStoreService 실투자금액 계산 수정
- PortfolioService 리셋 순서 안전하게 변경
- 랭킹 페이지 목업 50명 제거 → 실제 유저 데이터만 표시
- 항로 상점 데모 데이터 제거 + 구독자 수 실데이터 기반 재계산

+++ backend/src/main/java/com/project/whalearc/market/dto/CandlestickResponse.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 @AllArgsConstructor
 public class CandlestickResponse {
     private long time;       // Unix timestamp (seconds)
-    private long open;
-    private long high;
-    private long low;
-    private long close;
+    private double open;
+    private double high;
+    private double low;
+    private double close;
     private double volume;
 }
